requirements.txt
numpy>=1.10.0
spectral==0.19
scipy>=0.19.0
tqdm>=4.15.0
visdom>=0.1.5
seaborn>=0.8
scikit-learn>=0.19.0
scikit-image>=0.13.1
torch>=0.4.0
matplotlib>=2.0.2
torchsummary>=1.5
joblib
pytorch-ignite
torchinfo
spectral
